#743: Move editing of tournament details to dedicated component
